+++ js/tabitem/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gradio/tabitem
 
+## 0.8.3
+
+### Fixes
+
+- [#13644](https://github.com/gradio-app/gradio/pull/13644) [`f0abb1e`](https://github.com/gradio-app/gradio/commit/f0abb1ea8e61f4b886296365639b9fe0492012a9) - restore shared props lost during frontend prop partitioning.  Thanks @dawoodkhan82!
+
+### Dependency updates
+
+- @gradio/utils@0.14.0
+- @gradio/column@0.4.1
+- @gradio/tabs@0.10.0
+
 ## 0.8.2
 
 ### Dependency updates